v2-aiogram/backend/requirements.txt
aiogram
sqlalchemy 
aiosqlite
python-dotenv
fastapi
uvicorn
